Refactor FilterCommand into an abstract class that specific filters can inherit from
